requirements.txt
pydantic-ai[examples]>=0.0.8
python-dotenv>=1.0.0
aiohttp>=3.8.0
beautifulsoup4>=4.9.3
pytest>=8.0.0
trio>=0.27.0
logfire>=2.6.0
rich>=13.9.0